+++ GeneratedForm/index.jsx
@@ -1,159 +1,11 @@
 import React, {PropTypes, createElement, Component} from 'react';
-import {Field, FieldArray, reduxForm} from 'redux-form';
+import {reduxForm} from 'redux-form';
 import isString from 'lodash/isString';
 import isObject from 'lodash/isObject';
 import isFunction from 'lodash/isFunction';
-import isEqual from 'lodash/isEqual';
 
-export class DefaultFieldRow extends Component {
-	static propTypes = {
-		label: PropTypes.oneOfType([
-			PropTypes.string,
-			PropTypes.object,
-			PropTypes.func,
-			PropTypes.element
-		]).isRequired,
-		fieldProps: PropTypes.object.isRequired,
-		field: PropTypes.element.isRequired
-	};
-
-	render() {
-		let {label, fieldProps, field, ...props} = this.props;
-		return <div {...props}>
-			<label>{label}</label>
-			{field}
-		</div>;
-	}
-}
-
-class StaticElement extends Component {
-	static contextTypes = {
-		_reduxForm: PropTypes.object
-	};
-
-	render() {
-		let {component, ...props} = this.props;
-		return createElement(component, {
-			...props, formProps: this.context._reduxForm
-		});
-	}
-}
-
-// this prevents multiple triggers of the render function as we allways return the same function
-let cachedComponentFunctions = {};
-
-const buildFieldComponent = (field, type, CustomFieldComponent, formProps, rowComponent) => {
-	if (cachedComponentFunctions[`${formProps.formName}-${field.name}`] &&
-		isEqual(cachedComponentFunctions[`${formProps.formName}-${field.name}`].field, field) &&
-		cachedComponentFunctions[`${formProps.formName}-${field.name}`].type === type &&
-		cachedComponentFunctions[`${formProps.formName}-${field.name}`].CustomFieldComponent === CustomFieldComponent &&
-		cachedComponentFunctions[`${formProps.formName}-${field.name}`].rowComponent === rowComponent
-	) {
-		return cachedComponentFunctions[`${formProps.formName}-${field.name}`].component;
-	} else {
-		return (cachedComponentFunctions[`${formProps.formName}-${field.name}`] = {
-			field,
-			type,
-			CustomFieldComponent,
-			rowComponent,
-			component: fieldProps => {
-				let _inputField;
-				let {input, meta, ...props} = fieldProps;
-				if (!CustomFieldComponent) {
-					switch (type) {
-						case 'select':
-							let {optionValue, optionDisplay, ...settings} = (field.settings || {});
-							_inputField = <select {...input} {...props} {...settings}>{field.options.map((option, index) => {
-								if (isObject(option)) {
-									return <option key={index}
-												   value={option[optionValue || 'value']}>{option[optionDisplay || 'label']}</option>;
-								} else {
-									return <option key={index} value={option}>{option}</option>;
-								}
-							})}</select>;
-							break;
-						case 'textarea':
-							_inputField = <textarea {...input} {...props} {...field.settings}/>;
-							break;
-						default:
-							_inputField = <input {...input} {...props} {...field.settings} type={type || 'text'}/>;
-							break;
-					}
-				} else {
-					_inputField =
-						<CustomFieldComponent {...field} {...input} {...props} meta={meta}/>;
-				}
-				return createElement(rowComponent, {
-					...(field.rowProps || {}),
-					label: field.label || field.displayName || field.name,
-					fieldProps,
-					field: _inputField
-				});
-			}
-		}).component;
-	}
-}
-
-export class FormComponent extends Component {
-	static propTypes = {
-		formName: PropTypes.string.isRequired,
-		fieldsDefinition: PropTypes.array.isRequired,
-		rowComponent: PropTypes.oneOfType([PropTypes.func, PropTypes.string]),
-		formComponent: PropTypes.oneOfType([PropTypes.func, PropTypes.string])
-	};
-
-	static contextTypes = {
-		customFields: PropTypes.object,
-		defaultRowComponent: PropTypes.oneOfType([PropTypes.func, PropTypes.string]),
-		defaultFormComponent: PropTypes.oneOfType([PropTypes.func, PropTypes.string])
-	};
-
-	render() {
-		let {
-			rowComponent = this.context.defaultRowComponent || DefaultFieldRow,
-			formComponent = this.context.defaultFormComponent || 'form',
-			onSubmit,
-			handleSubmit,
-			..._formProps
-		} = this.props;
-		let {fieldsDefinition, ...formProps} = _formProps;
-		if (isString(formComponent)) {
-			formProps = {
-				style: formProps.style,
-				className: formProps.className,
-				method: formProps.method,
-				action: formProps.action
-			};
-		}
-		formProps = {
-			...formProps,
-			onSubmit: (onSubmit) ? handleSubmit(onSubmit) : undefined,
-			name: _formProps.formName
-		};
-		return createElement(formComponent, formProps, ..._formProps.fieldsDefinition.map((origField, index) => {
-			let {type, required, nonInteractive, Custom, ...field} = origField;
-			let ReduxFieldElement = Field;
-			let CustomFieldComponent;
-			if (isString(type)) {
-				CustomFieldComponent = (this.context.customFields || {})[type];
-			} else if(Custom) {
-				return <Custom key={index} field={origField}/>;
-			}
-			if (required) {
-				field.settings = {
-					...(field.settings || {}),
-					required
-				};
-			}
-			if (CustomFieldComponent && CustomFieldComponent.isArrayField) {
-				ReduxFieldElement = FieldArray;
-			} else if (nonInteractive || (CustomFieldComponent && CustomFieldComponent.nonInteractive)) {
-				ReduxFieldElement = StaticElement;
-			}
-			return <ReduxFieldElement key={index} name={field.name} component={buildFieldComponent(field, type, CustomFieldComponent, formProps, rowComponent)}/>;
-		}))
-	};
-}
+import DefaultFieldRow from './DefaultFieldRow'
+import FormComponent from './FormComponent';
 
 const ReduxForm = reduxForm()(FormComponent);
 
@@ -197,10 +49,36 @@ export default class GeneratedForm extends Component {
 	};
 
 	static contextTypes = {
-		buildErrorMessage: PropTypes.func,
-		formValidators: PropTypes.object
+		generatedForm: PropTypes.object
 	};
-
+	
+	// THe following creates the defaults for the generatedForm context
+	static childContextTypes = {
+		generatedForm: PropTypes.object
+	};
+	
+	getChildContext() {
+		let _generatedForm = this.context.generatedForm || {};
+		let generatedForm = {
+			defaultFormComponent: 'form',
+			defaultRowComponent: DefaultFieldRow,
+			objectArray: {
+				arrayComponent: (({fields, fieldsMap, displayName, meta, ...props}) => <div {...props}>
+					{fields.map(fieldsMap)}
+					<button type="button" onClick={() => fields.push({})}>Add</button>
+				</div>),
+				arrayRowComponent: (({index, formName, children, fields}) => <div>
+					{children}
+					<button type="button" onClick={() => fields.remove(index)}>remove</button>
+				</div>),
+				arrayFormRowComponent: _generatedForm.defaultRowComponent || DefaultFieldRow,
+				..._generatedForm.objectArray
+			},
+			..._generatedForm
+		};
+		return {generatedForm};
+	}
+	
 	shouldComponentUpdate(nextProps) {
 		return this.props.formName !== nextProps.formName || this.props.fieldsDefinition !== nextProps.fieldsDefinition || this.props.formRedux !== nextProps.formRedux;
 	}
@@ -211,7 +89,7 @@ export default class GeneratedForm extends Component {
 			if (field.validator) {
 				field.validators = [field.validator];
 			}
-			let contextValidators = this.context.formValidators || {
+			let contextValidators = (this.context.generatedForm || {}).formValidators || {
 					required: _requiredValidator,
 					email: _emailValidator
 				};