+++ Stateless/PanelHeader.jsx
@@ -2,6 +2,9 @@ import React, {Component} from 'react';
 import PanelHeader from 'rebass/dist/PanelHeader';
 
 function WrappedPanelHeader({style = {}, theme = 'primary', ...props}, {rebass}) {
+	if (props.inverted || theme === 'default') {
+		style.color = rebass.colors.black;
+	}
 	return <PanelHeader
 		style={{
 			...style,